+++ pysui/sui/sui_config.py
@@ -21,7 +21,7 @@ from io import TextIOWrapper
 from pathlib import Path
 import json
 import yaml
-from deprecated.sphinx import deprecated, versionadded
+from deprecated.sphinx import versionadded
 from pysui.abstracts import ClientConfiguration, SignatureScheme, KeyPair
 from pysui.sui.sui_constants import (
     PYSUI_EXEC_ENV,
@@ -30,10 +30,8 @@ from pysui.sui.sui_constants import (
     DEFAULT_DEVNET_PATH_STRING,
     DEVNET_FAUCET_URL,
     DEVNET_SOCKET_URL,
-    LOCALNET_ENVIRONMENT_KEY,
     LOCALNET_FAUCET_URL,
     LOCALNET_SOCKET_URL,
-    TESTNET_ENVIRONMENT_KEY,
     TESTNET_FAUCET_URL,
     TESTNET_SOCKET_URL,
 )
@@ -65,45 +63,6 @@ def _set_env_vars(client_config_path: Path, sui_exec_path: Path):
 class SuiConfig(ClientConfiguration):
     """Sui default configuration class."""
 
-    def _old_init(self, config_path: str, env: str, active_address: str, keystore_file: str, current_url: str) -> None:
-        """__init__ SuiConfig initialization.
-
-        :param config_path: Fully qualified path to client.yaml configuration to use.
-        :type config_path: str
-        :param env: The active environment name (e.g. devnet, localnet)
-        :type env: str
-        :param active_address: Which address to set as active-address
-        :type active_address: str
-        :param keystore_file: Fully qualifed path to keystore file
-        :type keystore_file: str
-        :param current_url: URL of SUI gateway to use (e.g. https://fullnode.devnet.sui.io:443)
-        :type current_url: str
-        :raises SuiInvalidKeystringLength: If, when ingesting keys, pre 0.21.0 keystring found
-        :raises SuiNoKeyPairs: If keystore file is empty
-        :raises SuiKeystoreFileError: If exception occured during keystore file read
-        :raises SuiKeystoreAddressError: If JSON error reading keystring array from keystore file
-        :raises SuiFileNotFound: If path to keystore file does not exist
-        """
-        super().__init__(config_path, keystore_file)
-        self._active_address = SuiAddress(active_address)
-        self._current_url = current_url
-        self._current_env = env
-        # Add for foward capability
-        _set_env_vars(config_path, Path(os.path.expanduser(DEFAULT_SUI_BINARY_PATH)))
-        if env == LOCALNET_ENVIRONMENT_KEY:
-            self._faucet_url = LOCALNET_FAUCET_URL
-            self._socket_url = LOCALNET_SOCKET_URL
-            self._local_running = True
-        elif env == TESTNET_ENVIRONMENT_KEY:
-            self._faucet_url = TESTNET_FAUCET_URL
-            self._socket_url = TESTNET_SOCKET_URL
-            self._local_running = False
-        else:
-            self._faucet_url = DEVNET_FAUCET_URL
-            self._socket_url = DEVNET_SOCKET_URL
-            self._local_running = False
-        self._keypairs, self._addresses, self._address_keypair = load_keys_and_addresses(keystore_file)
-
     def _initiate(self, active_address: str, rpc_url: str, environment: str) -> None:
         """."""
         self._active_address = SuiAddress(active_address)
@@ -124,10 +83,6 @@ class SuiConfig(ClientConfiguration):
             case "mainnet":
                 raise NotImplementedError("mainnet not deployed for Sui network yet.")
         self._keypairs, self._addresses, self._address_keypair = load_keys_and_addresses(self.keystore_file)
-
-    @deprecated(version="0.16.1", reason="To support more robust configurations such as sui-base")
-    def _d_init__(self, config_path: str, env: str, active_address: str, keystore_file: str, current_url: str) -> None:
-        """_d_init__ old init function."""
 
     def _write_keypair(self, keypair: KeyPair, file_path: str = None) -> None:
         """Register the keypair and write out to keystore file."""
@@ -190,44 +145,6 @@ class SuiConfig(ClientConfiguration):
         return (str(fpath), active_env, active_address, keystore_file, current_url)
 
     @classmethod
-    @deprecated(version="0.16.0", reason="Accomodate more flexible setup use default_config or sui_base instead.")
-    def default(cls) -> "SuiConfig":
-        """default Looks for and loads client.yaml from ~/.sui/sui_config.
-
-        :raises SuiFileNotFound: If client.yaml file not found in default path
-        :return: An instance of SuiConfig
-        :rtype: SuiConfig
-        """
-        expanded_path = os.path.expanduser(DEFAULT_DEVNET_PATH_STRING)
-        if os.path.exists(expanded_path):
-            with open(expanded_path, encoding="utf8") as core_file:
-                config = super(ClientConfiguration, cls).__new__(cls)
-                config._old_init(*cls._parse_config(Path(expanded_path), core_file))
-                return config
-        else:
-            raise SuiFileNotFound(f"{expanded_path} not found.")
-
-    @classmethod
-    @deprecated(version="0.16.1", reason="Removing in favor of default or sui_base")
-    def from_config_file(cls, infile: str) -> "SuiConfig":
-        """from_config_file Load a SuiConfig from a fully qualified path to client.yaml.
-
-        :param infile: Path to client.yaml file to load
-        :type infile: str
-        :raises SuiFileNotFound: If client.yaml does not exist in path provided
-        :return: An instance of SuiConfig
-        :rtype: SuiConfig
-        """
-        expanded_path = os.path.expanduser(infile)
-        if os.path.exists(expanded_path):
-            with open(expanded_path, encoding="utf8") as core_file:
-                config = super(ClientConfiguration, cls).__new__(cls)
-                config._old_init(*cls._parse_config(Path(expanded_path), core_file))
-                return config
-        else:
-            raise SuiFileNotFound(f"{expanded_path} not found.")
-
-    @classmethod
     def _new_parse_config(cls, sui_config: str) -> tuple[str, str, str]:
         """New Config Parser."""
         active_address = sui_config["active_address"] if "active_address" in sui_config else None
@@ -271,12 +188,6 @@ class SuiConfig(ClientConfiguration):
     def sui_base_config(cls) -> "SuiConfig":
         """."""
         return cls._create_config(*sui_base_get_config())
-
-    @classmethod
-    @deprecated(version="0.16.1", reason="Never implemented. Will be removed in next version.")
-    def _generate_configuration(cls) -> "ClientConfiguration":
-        """Generate a default configuration."""
-        raise NotImplementedError("SuiConfig.generate_configuration not implemented yet.")
 
     @property
     def rpc_url(self) -> str: